Move launch functions to controller folder.
